Server side database implementation
A database is created in the backend with mangoose.
The database contains 4 tables: User, Project, Discussion, Message.

Every message will be stored in the database.
This will be in use when implementing the project lead use case, in which
they can see all the messages in a specific project.

For now admins and project leads are defined statically through the users.json file.