New panel inside extension + data persistency
A new panel is created inside the vscode extension, that will in time host
the webview for the project lead interface.

Extension data persistency is being implemented. For now the access token
is stored in the extension's global state. New backend route and function
was added to check the validity of a token.